test(auth): add other tests for nsec
